requirements.txt
Django==1.6.1
South==0.8.4
dj-static==0.0.5
django-debug-toolbar==1.0
django-extensions==1.3.3
django-pin-passcode>=0.1.1
gunicorn==18.0
ipdb==0.8
ipython==1.1.0
psycopg2==2.5.1
pyreadline==2.0
python-dateutil==2.2
raven==4.2.1
six==1.5.2
sqlparse==0.1.10
static==0.4
wsgiref==0.1.2
django-qsstats-magic
numpy==1.8.1
